Rework user login flow
