+++ src/components/Home/Home.js
@@ -1,33 +1,112 @@
-import React from 'react';
+import React, { Component } from 'react';
 import axios from 'axios'
 import { Link } from 'react-router-dom'
 
-const Home = (props) => {
-  
-  const handleClick = () => {
-    axios.delete('http://localhost:3005/logout', {withCredentials: true})
-    .then(response => {
-      props.handleLogout()
-      props.history.push('/')
-    })
-    .catch(error => console.log(error))
+class Home extends Component {
+  constructor(props) {
+    super(props);
+    this.state = { 
+      username: '',
+      email: '',
+      password: '',
+      errors: ''
+     };
+  }
+  componentWillMount() {
+    return this.props.loggedInStatus ? this.redirect() : null
   }
 
-  return (
-   
-    <div>
-      <Link to='/login'>Log In</Link>
-      <br></br>
-      <Link to='/signup'>Sign Up</Link>
-      <br></br>
-      { 
-        props.loggedInStatus ? 
-        <Link to='/logout' onClick={handleClick}>Log Out</Link> : 
-        null
+  handleChange = (event) => {
+    const {name, value} = event.target
+    this.setState({
+      [name]: value
+    })
+  };
+
+  handleSubmit = (event) => {
+    event.preventDefault()
+    const {username, email, password} = this.state
+    let user = {
+      username: username,
+      email: email,
+      password: password
+    }
+    
+    axios.post('http://localhost:3005/login', {user}, {withCredentials: true})
+    .then(response => {
+      if (response.data.logged_in) {
+        this.props.handleLogin(response.data)
+        this.redirect()
+      } else {
+        this.setState({
+          errors: response.data.errors
+        })
       }
-    </div>
-  );
+    })
+    .catch(error => console.log('api errors:', error))
+  };
 
-};
+  redirect = () => {
+    this.props.history.push('/dashboard')
+  }
 
+  handleErrors = () => {
+    return (
+      <div>
+        <ul>
+        {this.state.errors.map(error => {
+        return <li key={error}>{error}</li>
+          })
+        }
+        </ul>
+      </div>
+    )
+  }
+  
+  render() {
+    const {username, email, password} = this.state
+
+    return (
+      <div>
+        <h1>Log In</h1>
+        <form onSubmit={this.handleSubmit}>
+          <input
+            placeholder="username"
+            type="text"
+            name="username"
+            value={username}
+            onChange={this.handleChange}
+          />
+          <input
+            placeholder="email"
+            type="text"
+            name="email"
+            value={email}
+            onChange={this.handleChange}
+          />
+          <input
+            placeholder="password"
+            type="password"
+            name="password"
+            value={password}
+            onChange={this.handleChange}
+          />
+          <button placeholder="submit" type="submit">
+            Log In
+          </button>
+          <div>
+            or <Link to='/signup'>sign up</Link>
+          </div>
+          
+          </form>
+          <div>
+          {
+            this.state.errors ? this.handleErrors() : null
+          }
+        </div>
+      </div>
+    );
+  }
+
+}
 export default Home;
